rename src/Blog to src/Statie
